refactor: major codebase cleanup and model-compare database migration
## Database Changes
### Added Model-Compare Source Type
- Created migration 026-add-model-compare-source.sql to add 'model-compare' as valid conversation source
- Updated model-compare component to use 'model-compare' source instead of 'chat'
- This prevents model comparison conversations from appearing in the main chat conversation list
- NOTE: Database migration needs manual deployment via CDK

### Removed Duplicate Components (Reduced ~500 lines)
- Deleted /components/ui/message.tsx (35 lines) - using chat version
- Deleted /components/ui/chat-input.tsx (88 lines) - using chat version
- Deleted /components/utilities/error-boundary.tsx (57 lines) - using main version
- Deleted /app/(protected)/chat/_components/model-selector.tsx (134 lines) - using features version
- Updated all imports to use the canonical versions

### Consolidated Layout Files
- Created StandardPageLayout component to eliminate duplicate layout code
- Replaced identical layouts in dashboard, ideas, and utilities/assistant-architect
- Removed empty meta-prompting layout file
- Reduced layout duplication from 5 files to 1 shared component

### Fixed Type Issues
- Added placeholder prop support to ChatInput component
- Fixed ErrorBoundary import to use named export
- Fixed Message component props in assistant-architect-chat
- Removed invalid fallbackMessage prop from ErrorBoundary usage
- All components now pass TypeScript type checking

### Code Quality Improvements
- All changes pass ESLint with zero errors
- All changes pass TypeScript type checking
- Removed obsolete 025-enhanced-message-tracking.sql reference from db-init handler
- Updated all component imports to use consistent paths

## Benefits
- **Reduced bundle size** by ~500 lines through component deduplication
- **Improved maintainability** with single source of truth for components
- **Better separation** of model-compare and chat conversations
- **Cleaner architecture** with shared layout patterns
- **Type safety** ensured across all modified components

## Testing Status
- ✅ Linting passes with zero errors
- ✅ TypeScript type checking passes
- ✅ All component imports resolved correctly
- ⚠️ Database migration requires manual CDK deployment
