+++ app/(protected)/dashboard/layout.tsx
@@ -1,18 +1,9 @@
-import { NavbarNested } from "@/components/navigation/navbar-nested"
+import StandardPageLayout from "@/components/layouts/standard-page-layout"
 
 export default function DashboardLayout({
   children
 }: {
   children: React.ReactNode
 }) {
-  return (
-    <div className="flex min-h-screen pt-14">
-      <NavbarNested />
-      <main className="flex-1 lg:pl-[68px]">
-        <div className="bg-white p-4 sm:p-6 md:p-8">
-          {children}
-        </div>
-      </main>
-    </div>
-  )
+  return <StandardPageLayout>{children}</StandardPageLayout>
 }
