--- app/(protected)/meta-prompting/layout.tsx
@@ -1,7 +1,0 @@
-export default async function MetaPromptingLayout({
-  children
-}: {
-  children: React.ReactNode
-}) {
-  return children
-}
