+++ components/layouts/standard-page-layout.tsx
@@ -1,0 +1,20 @@
+import { NavbarNested } from "@/components/navigation/navbar-nested"
+
+interface StandardPageLayoutProps {
+  children: React.ReactNode
+}
+
+export default function StandardPageLayout({
+  children
+}: StandardPageLayoutProps) {
+  return (
+    <div className="flex min-h-screen pt-14">
+      <NavbarNested />
+      <main className="flex-1 lg:pl-[68px]">
+        <div className="bg-white p-4 sm:p-6 md:p-8">
+          {children}
+        </div>
+      </main>
+    </div>
+  )
+}
