+++ infra/database/lambda/db-init-handler.ts
@@ -44,7 +44,7 @@ const MIGRATION_FILES = [
   '020-add-user-role-version.sql',
   '023-navigation-multi-roles.sql',
   '024-model-role-restrictions.sql',
-  '025-enhanced-message-tracking.sql'
+  '026-add-model-compare-source.sql'
   // ADD NEW MIGRATIONS HERE - they will run once and be tracked
 ];
